CA/intelligent_invoice_processor/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pillow>=8.2.0
pytesseract>=0.3.8
pdf2image>=1.16.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2  # For improved fuzzywuzzy performance
opencv-python>=4.5.3
pytest>=6.2.5